wip: format lint fix

+++ pkg/util/lowlevel/pci.go
@@ -33,6 +33,7 @@ package lowlevel
 import (
 	"C"
 )
+
 import (
 	"fmt"
 	"sync"
@@ -165,21 +166,21 @@ func ReadLong(dev *PCIDev, addr int32) uint32 {
 }
 
 func ReadSMNApp(dev *PCIDev, addr uint32) uint32 {
-	WriteLong(dev, IOHC_NB_SMN_INDEX_0_REG, uint32(addr))
+	WriteLong(dev, IOHC_NB_SMN_INDEX_0_REG, addr)
 	return ReadLong(dev, IOHC_NB_SMN_DATA_0_REG)
 }
 
 func WriteSMNApp(dev *PCIDev, addr, data uint32) {
-	WriteLong(dev, IOHC_NB_SMN_INDEX_0_REG, uint32(addr))
+	WriteLong(dev, IOHC_NB_SMN_INDEX_0_REG, addr)
 	WriteLong(dev, IOHC_NB_SMN_DATA_0_REG, data)
 }
 
 func ReadSMN(dev *PCIDev, addr uint32) uint32 {
-	WriteLong(dev, IOHC_NB_SMN_INDEX_3_REG, uint32(addr))
+	WriteLong(dev, IOHC_NB_SMN_INDEX_3_REG, addr)
 	return ReadLong(dev, IOHC_NB_SMN_DATA_3_REG)
 }
 
 func WriteSMN(dev *PCIDev, addr, data uint32) {
-	WriteLong(dev, IOHC_NB_SMN_INDEX_3_REG, uint32(addr))
+	WriteLong(dev, IOHC_NB_SMN_INDEX_3_REG, addr)
 	WriteLong(dev, IOHC_NB_SMN_DATA_3_REG, data)
 }
